+++ receiver.h
@@ -19,9 +19,9 @@ storage_handle receiver_get_storage(receiver_handle recv);
 boolean receiver_is_running(receiver_handle recv);
 status receiver_stop(receiver_handle recv);
 
-long receiver_get_tcp_gap_count(receiver_handle recv);
-long receiver_get_tcp_bytes_recv(receiver_handle recv);
-long receiver_get_mcast_bytes_recv(receiver_handle recv);
+size_t receiver_get_tcp_gap_count(receiver_handle recv);
+size_t receiver_get_tcp_bytes_recv(receiver_handle recv);
+size_t receiver_get_mcast_bytes_recv(receiver_handle recv);
 
 #ifdef __cplusplus
 }
