+++ sender.h
@@ -21,9 +21,9 @@ status sender_flush(sender_handle send);
 boolean sender_is_running(sender_handle send);
 status sender_stop(sender_handle send);
 
-long sender_get_tcp_gap_count(sender_handle send);
-long sender_get_tcp_bytes_sent(sender_handle send);
-long sender_get_mcast_bytes_sent(sender_handle send);
+size_t sender_get_tcp_gap_count(sender_handle send);
+size_t sender_get_tcp_bytes_sent(sender_handle send);
+size_t sender_get_mcast_bytes_sent(sender_handle send);
 int sender_get_subscriber_count(sender_handle send);
 
 #ifdef __cplusplus
